reimplemented DatabaseManager::createDatabase()
